Split up UONET status utility functions
